perform delete thing

+++ controllers/stuff.controller.js
@@ -1,4 +1,5 @@
 const Thing = require("../models/thing.model");
+const fs = require("fs");
 
 // exports.createThing = (req, res, next) => {
 //   const thing = new Thing({
@@ -109,17 +110,38 @@ exports.modifyThing = (req, res, next) => {
     });
 };
 
+// exports.deleteThing = (req, res, next) => {
+//   Thing.deleteOne({ _id: req.params.id })
+//     .then(() => {
+//       res.status(200).json({
+//         message: "Deleted!",
+//       });
+//     })
+//     .catch((error) => {
+//       res.status(400).json({
+//         error: error,
+//       });
+//     });
+// };
+
 exports.deleteThing = (req, res, next) => {
-  Thing.deleteOne({ _id: req.params.id })
-    .then(() => {
-      res.status(200).json({
-        message: "Deleted!",
-      });
+  Thing.findOne({ _id: req.params.id })
+    .then((thing) => {
+      if (thing.userId != req.auth.userId) {
+        res.status(401).json({ message: "Not authorized" });
+      } else {
+        const filename = thing.imageUrl.split("/images/")[1];
+        fs.unlink(`images/${filename}`, () => {
+          Thing.deleteOne({ _id: req.params.id })
+            .then(() => {
+              res.status(200).json({ message: "Objet supprimé !" });
+            })
+            .catch((error) => res.status(401).json({ error }));
+        });
+      }
     })
     .catch((error) => {
-      res.status(400).json({
-        error: error,
-      });
+      res.status(500).json({ error });
     });
 };
